% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgd2r-package.R
\docType{package}
\name{fgd2r-package}
\alias{fgd2r}
\alias{fgd2r-package}
\title{fgd2r: Reading Fundamental Geospatial Data}
\description{
This package enables to read fundamental geospatial data of GSI Japan, such as DEM, Landform, etc.
}
\author{
\strong{Maintainer}: Kiwamu Ishikura \email{ishikura.kiwamu@gmail.com}

}
