% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.R
\name{read_fgd_dem}
\alias{read_fgd_dem}
\title{read_fgd_dem}
\usage{
read_fgd_dem(file_path)
}
\arguments{
\item{file_path}{A file path to read (xml or zip).}
}
\value{
An object of \code{SpatRaster}.
}
\description{
To read FGD DEM file.
}
\examples{
dem_644320 <- system.file("extdata", "FG-GML-6443-20-DEM5A.zip", package = "fgd2r")
trr_644320 <- read_fgd_dem(dem_644320)
}
